# plugins/mnemosyne/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Mnemosyne Knowledge Management API
  description: |
    The Mnemosyne API provides comprehensive knowledge management capabilities including:
    - Knowledge node creation and management
    - Relationship mapping between nodes
    - Advanced search functionality
    - Graph visualization data
    - Import/export capabilities
    - Template management
  version: 1.0.0
  contact:
    name: Alexandria Platform Team
    email: support@alexandria.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/mnemosyne
    description: Development server
  - url: https://api.alexandria.dev/mnemosyne
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /relationships:
    get:
      tags:
        - Relationships
      summary: List relationships
      description: Get all relationships with optional filtering
      operationId: listRelationships
      parameters:
        - name: nodeId
          in: query
          description: Filter by source or target node ID
          schema:
            type: string
        - name: type
          in: query
          description: Filter by relationship type
          schema:
            type: string
            enum: [links-to, references, contains, tagged-with, derived-from]
        - name: direction
          in: query
          description: Direction of relationships (when nodeId is provided)
          schema:
            type: string
            enum: [incoming, outgoing, both]
            default: both
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Relationship'

    post:
      tags:
        - Relationships
      summary: Create relationship
      description: Create a new relationship between nodes
      operationId: createRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRelationshipInput'
      responses:
        '201':
          description: Relationship created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Relationship'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Relationship:
      type: object
      required:
        - id
        - sourceId
        - targetId
        - type
      properties:
        id:
          type: string
        sourceId:
          type: string
        targetId:
          type: string
        type:
          type: string
          enum: [links-to, references, contains, tagged-with, derived-from]
        metadata:
          type: object
        created:
          type: string
          format: date-time
    CreateRelationshipInput:
      type: object
      required:
        - sourceId
        - targetId
        - type
      properties:
        sourceId:
          type: string
        targetId:
          type: string
        type:
          type: string
          enum: [links-to, references, contains, tagged-with, derived-from]
        metadata:
          type: object
...